requirements.txt
pytest~=7.4.4
selene~=2.0.0rc6
allure-pytest==2.13.0
python-dotenv==1.0.1
requests

